% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scala_list}
\alias{scala_list}
\title{Generate a Scala List}
\usage{
scala_list(sc, x)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{x}{A vector.}
}
\description{
Scala List cannot be generated using sparklyr. We must generate them using
this function.
}
\examples{
sc <- sparklyr::spark_connect(master = "local", version = "2.1.0")
sparkts:::scala_list(sc, c(1, 2, 3))
sparklyr::spark_disconnect(sc = sc)
}
