% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scala_seq}
\alias{scala_seq}
\title{Generate a Scala Seq}
\usage{
scala_seq(sc, x)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{x}{A vector.}
}
\description{
Scala Seq (sequence) cannot be generated using sparklyr. We must generate
them using this function.
}
\examples{
sc <- sparklyr::spark_connect(master = "local", version = "2.1.0")
sparkts:::scala_seq(sc, c(1, 2, 3))
sparklyr::spark_disconnect(sc = sc)
}
