% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_standard_error.R
\name{sdf_standard_error_r6}
\alias{sdf_standard_error_r6}
\title{Calculate the standard error}
\description{
This function will add an extra column on to a Spark DataFrame containing the
standard error.
}
\section{Usage}{

\preformatted{
  p <- sdf_standard_error_r6$new(sc = sc, data = std_data)
  p$standard_error(
    x_col = "xColumn", y_col = "yColumn", z_col = "zColumn",
    new_column_name = "test"
  )
}
}

\section{Arguments}{

\describe{
  \item{p}{A \code{waterflow} object.}
  \item{wall}{A numeric vector of wall heights}
}
}

\section{Details}{

\code{$new()} instantiates the class.

\code{$standard_error()} calculates the standard error and returns a
  \code{data.frame}.
}

\examples{
# Set up a spark connection
sc <- sparklyr::spark_connect(master = "local", version = "2.1.0")

# Extract some data
std_data <- sparklyr::spark_read_json(
  sc,
  "std_data",
  path = system.file(
    "data_raw/StandardErrorDataIn.json",
    package = "sparkts"
  )
) \%>\%
  sparklyr::spark_dataframe()

# Instantiate the class
p <- sdf_standard_error_r6(sc = sc, data = std_data)
# Calculate the standard errors
p$standard_error(
  x_col = "xColumn", y_col = "yColumn", z_col = "zColumn",
  new_column_name = "test"
)
}
