% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testsml-package.R
\docType{package}
\name{testsml-package}
\alias{testsml}
\alias{testsml-package}
\title{A test package for calling to spakr-ts package}
\description{
This package acts as a test bed for learning the sparklyr API
}
\author{
\strong{Maintainer}: Nathan Eastwood \email{nathan.eastwood@icloud.com}

}
