% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{utils}
\alias{utils}
\title{View the spark DataFrames}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
utils
}
\value{
A tibble
}
\description{
This function "collects" the DataFrame in the spark connection and returns it
as a tibble in R
}
\keyword{datasets}
