% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark_standard_error.R
\name{sdf_standard_error}
\alias{sdf_standard_error}
\title{Calculate the standard error}
\usage{
sdf_standard_error(sc, data, x_col, y_col, z_col, new_column_name)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{data}{The Spark \code{DataFrame} on which to perform the function.}

\item{x_col}{The column to be used as X in the calculation.}

\item{y_col}{The column to be used as Y in the calculation.}

\item{z_col}{The column to be used as Z in the calculation.}

\item{new_column_name}{The name of the standard error column.}
}
\description{
This function will add an extra column on to a Spark DataFrame containing the
standard error.
}
\examples{
# Set up a spark connection
sc <- sparklyr::spark_connect(master = "local", version = "2.1.0")

# Extract some data
std_data <- sparklyr::spark_read_json(
  sc,
  "std_data",
  path = system.file(
    "data_raw/StandardErrorDataIn.json",
    package = "testsml"
  )
) \%>\%
  sparklyr::spark_dataframe()

# Calculate the standard error
output <- sdf_standard_error(
  sc,
  std_data,
  x_col = "xColumn",
  y_col = "yColumn",
  z_col = "zColumn",
  new_column_name = "stdError"
)

# Extract the standard error column
output <- output \%>\% dplyr::collect()
output[["stdError"]]

sparklyr::spark_disconnect(sc = sc)

}
