% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkts-package.R
\docType{package}
\name{sparkts-package}
\alias{sparkts}
\alias{sparkts-package}
\title{A package for calling Scala methods from the spark-ts package}
\description{
A package for calling Scala methods from the spark-ts package
}
\author{
\strong{Maintainer}: Nathan Eastwood \email{nathan.eastwood@icloud.com}

}
